Extract hashing stuff out of Language.h into LanguageHash.h
